pyproject.toml
[project]
name = "manifest-quiz"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "nasuka", email = "nsk.smn@gmail.com" }
]
dependencies = [
    "streamlit>=1.45.1",
    "pandas>=2.0.0",
    "openrouter>=1.0",
    "python-dotenv>=1.1.0",
]
readme = "README.md"
requires-python = ">= 3.12"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/manifest_quiz"]
